- split home screen to components
